requirements.txt
discord.py
requests
bs4
requests
discord-py-slash-command